+++ .expo/packager-info.json
@@ -1,4 +1,9 @@
 {
   "expoServerPort": 8081,
-  "packagerPort": 8081
+  "packagerPort": 8081,
+  "packagerPid": null,
+  "expoServerNgrokUrl": null,
+  "packagerNgrokUrl": null,
+  "ngrokPid": null,
+  "webpackServerPort": null
 }
